fixed module hierarchy
