+++ src-tauri/src/algorithms/mod.rs
@@ -1,0 +1,2 @@
+mod articulation_point;
+mod globalmincut;
